api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
psycopg2-binary>=2.9.9
pydantic>=2.6.0
python-multipart>=0.0.9
httpx>=0.26.0
